removed unnecessary RoleService, RoleServiceImpl, 05-insert-books.yaml, again renamed UserDetailsService -> CustomUserDetailsService | add  GrantedAuthority implementation for Role GrantedAuthority | and some minor fixes
