+++ src/main/java/dev/vasyl/proj/config/SecurityConfig.java
@@ -2,7 +2,7 @@ package dev.vasyl.proj.config;
 
 import static org.springframework.security.config.Customizer.withDefaults;
 
-import dev.vasyl.proj.sucurity.UserDetailsService;
+import dev.vasyl.proj.security.CustomUserDetailsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,7 +19,7 @@ import org.springframework.security.web.SecurityFilterChain;
 @Configuration
 public class SecurityConfig {
 
-    private final UserDetailsService userDetailsService;
+    private final CustomUserDetailsService customUserDetailsService;
 
     @Bean
     public PasswordEncoder getPasswordEncoder() {
@@ -45,7 +45,7 @@ public class SecurityConfig {
                 .httpBasic(withDefaults())
                 .sessionManagement(session ->
                         session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .userDetailsService(userDetailsService)
+                .userDetailsService(customUserDetailsService)
                 .build();
     }
 }
