+++ src/main/java/dev/vasyl/proj/repository/RoleRepository.java
@@ -4,7 +4,9 @@ import dev.vasyl.proj.model.Role;
 import dev.vasyl.proj.model.RoleName;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByName(RoleName roleName);
 }
