+++ src/main/java/dev/vasyl/proj/security/CustomUserDetailsService.java
@@ -1,0 +1,22 @@
+package dev.vasyl.proj.security;
+
+import dev.vasyl.proj.exception.EntityNotFoundException;
+import dev.vasyl.proj.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException(
+                        "Cant find user by email: [" + email + "]"));
+    }
+}
