--- src/main/java/dev/vasyl/proj/sucurity/UserDetailsService.java
@@ -1,41 +1,0 @@
-package dev.vasyl.proj.sucurity;
-
-import dev.vasyl.proj.exception.EntityNotFoundException;
-import dev.vasyl.proj.repository.UserRepository;
-import java.util.List;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsService
-        implements org.springframework.security.core.userdetails.UserDetailsService {
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email)
-                .map(user -> {
-                    List<GrantedAuthority> authorities = user.getRoles().stream()
-                            .map(role -> new SimpleGrantedAuthority("ROLE_" + role.getName()))
-                            .collect(Collectors.toList());
-                    if (authorities.contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
-                        authorities.addAll(List.of(
-                                new SimpleGrantedAuthority("ROLE_USER")
-                        ));
-                    }
-                    return new org.springframework.security.core.userdetails.User(
-                            user.getEmail(),
-                            user.getPassword(),
-                            authorities
-                    );
-                })
-                .orElseThrow(() -> new EntityNotFoundException(
-                        "Cant find user by email: [" + email + "]"));
-    }
-}
