--- src/main/resources/db/changelog/changes/05-insert-books.yaml
@@ -1,96 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: insert-books-1
-      author: vasyl
-      runOnChange: true
-      changes:
-        - preConditions:
-            onFail: MARK_RAN
-            sqlCheck:
-              expectedResult: 0
-              sql: SELECT COUNT(*) FROM books WHERE isbn = 'bookSomeIsbn1'
-        - insert:
-            tableName: books
-            columns:
-              - column:
-                  name: title
-                  value: "title1"
-              - column:
-                  name: author
-                  value: "bookAuthor1"
-              - column:
-                  name: isbn
-                  value: "bookSomeIsbn1"
-              - column:
-                  name: price
-                  value: 11
-              - column:
-                  name: description
-                  value: "bookDescription1"
-              - column:
-                  name: cover_image
-                  value: "bookCoverImage1"
-
-  - changeSet:
-      id: insert-books-2
-      author: vasyl
-      runOnChange: true
-      changes:
-        - preConditions:
-            onFail: MARK_RAN
-            sqlCheck:
-              expectedResult: 0
-              sql: SELECT COUNT(*) FROM books WHERE isbn = 'bookSomeIsbn2'
-        - insert:
-            tableName: books
-            columns:
-              - column:
-                  name: title
-                  value: "title2"
-              - column:
-                  name: author
-                  value: "bookAuthor2"
-              - column:
-                  name: isbn
-                  value: "bookSomeIsbn2"
-              - column:
-                  name: price
-                  value: 22
-              - column:
-                  name: description
-                  value: "bookDescription2"
-              - column:
-                  name: cover_image
-                  value: "bookCoverImage2"
-
-  - changeSet:
-      id: insert-books-3
-      author: vasyl
-      runOnChange: true
-      changes:
-        - preConditions:
-            onFail: MARK_RAN
-            sqlCheck:
-              expectedResult: 0
-              sql: SELECT COUNT(*) FROM books WHERE isbn = 'bookSomeIsbn3'
-        - insert:
-            tableName: books
-            columns:
-              - column:
-                  name: title
-                  value: "title3"
-              - column:
-                  name: author
-                  value: "bookAuthor3"
-              - column:
-                  name: isbn
-                  value: "bookSomeIsbn3"
-              - column:
-                  name: price
-                  value: 33
-              - column:
-                  name: description
-                  value: "bookDescription3"
-              - column:
-                  name: cover_image
-                  value: "bookCoverImage3"
