--- src/main/resources/db/changelog/changes/06-insert-roles.yaml
@@ -1,19 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: insert-roles
-      author: vasyl
-      runOnChange: true
-      changes:
-        - insert:
-            tableName: roles
-            columns:
-              - column:
-                  name: name
-                  value: USER
-
-        - insert:
-            tableName: roles
-            columns:
-              - column:
-                  name: name
-                  value: ADMIN
