+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -8,10 +8,8 @@ databaseChangeLog:
   - include:
       file: db/changelog/changes/04-create-user_roles-table.yaml
   - include:
-      file: db/changelog/changes/05-insert-books.yaml
+      file: db/changelog/changes/05-insert-roles.yaml
   - include:
-      file: db/changelog/changes/06-insert-roles.yaml
+      file: db/changelog/changes/06-insert-users.yaml
   - include:
-      file: db/changelog/changes/07-insert-users.yaml
-  - include:
-      file: db/changelog/changes/08-insert-admin-user.yaml
+      file: db/changelog/changes/07-insert-admin-user.yaml
